Updates code

+++ src/main/java/com/act/techtalk2022/controller/AttenderController.java
@@ -6,15 +6,18 @@ import com.act.techtalk2022.controller.response.AttenderResponse;
 import com.act.techtalk2022.controller.response.CreateAttenderResponse;
 import com.act.techtalk2022.controller.response.GeneralResponse;
 import com.act.techtalk2022.controller.response.GetAllAttenderResponse;
+import com.act.techtalk2022.exception.CommonException;
 import com.act.techtalk2022.factory.ResponseFactory;
 import com.act.techtalk2022.repository.enitiy.AttenderEntity;
 import com.act.techtalk2022.service.AttenderService;
+import com.act.techtalk2022.util.CommonUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Description;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -25,6 +28,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.time.Instant;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @RestController
@@ -63,6 +67,26 @@ public class AttenderController {
     }
 
     private void validateCreateAttender(CreateAttenderRequest request) {
+        if (null == request.getFullName() || request.getFullName().isEmpty()) {
+            log.error("The full name cannot be blank");
+
+            throw new CommonException(HttpStatus.BAD_REQUEST, "bad_request", "The full name cannot be blank");
+        }
+
+        if (null == request.getEmail() || request.getEmail().isEmpty()) {
+            log.error("The email cannot be blank or validate");
+
+            throw new CommonException(HttpStatus.BAD_REQUEST, "bad_request", "The email cannot be blank or validate");
+        }
+
+        if (null == request.getDateOfBirth()) {
+
+            log.error("The date of birth cannot be blank or validate");
+
+            throw new CommonException(HttpStatus.BAD_REQUEST, "bad_request", "The date of birth cannot be blank");
+
+        }
+        log.info("== End to validate user==");
     }
 
     //endregion Adds attender
@@ -112,6 +136,13 @@ public class AttenderController {
             @RequestBody UpdateAttenderRequest request) {
         log.info("========== Start to update an attender  ==========");
 
+        Optional<AttenderEntity> entity = attenderService.findById(attenderId);
+
+        if(!entity.isPresent()) {
+            log.error("The attender is not found");
+            throw new CommonException(HttpStatus.NOT_FOUND, "not_found", "The attender is not found");
+        }
+
         validateUpdateAttender(request);
         attenderService.updateAttender(attenderId, request);
 
@@ -120,6 +151,25 @@ public class AttenderController {
     }
 
     private void validateUpdateAttender(UpdateAttenderRequest request) {
+        if (null == request.getFullName() || request.getFullName().isEmpty()) {
+            log.error("The full name cannot be blank");
+
+            throw new CommonException(HttpStatus.BAD_REQUEST, "bad_request", "The full name cannot be blank");
+        }
+
+        if (null == request.getEmail() || request.getEmail().isEmpty()) {
+            log.error("The email cannot be blank or validate");
+
+            throw new CommonException(HttpStatus.BAD_REQUEST, "bad_request", "The email cannot be blank or validate");
+        }
+
+        if (null == request.getDateOfBirth()) {
+
+            log.error("The date of birth cannot be blank or validate");
+
+            throw new CommonException(HttpStatus.BAD_REQUEST, "bad_request", "The date of birth cannot be blank");
+
+        }
     }
 
     //endregion Update an attender
@@ -131,7 +181,13 @@ public class AttenderController {
             @PathVariable("id") Integer attenderId) {
         log.info("========== Start to delete an attender  ==========");
 
-        validateDeleteAttender(attenderId);
+        Optional<AttenderEntity> entity = attenderService.findById(attenderId);
+
+        if(!entity.isPresent()) {
+            log.error("The attender is not found");
+            throw new CommonException(HttpStatus.NOT_FOUND, "not_found", "The attender is not found");
+        }
+
 
         attenderService.deleteAttender(attenderId);
